Extract module test.on_yubikey.cli_piv.test_misc
